Completed styleSelector refactor, cleaned up some code in server index.js for deployment env
