+++ client/src/components/ProdDetail/ProdDetail.jsx
@@ -1,6 +1,5 @@
 /* eslint-disable camelcase */
 import React from 'react';
-// import { productTest, productStylesTest } from './testProduct';
 import PropTypes from 'prop-types';
 import ImageGallery from './ImageGallery';
 import StyleSelector from './StyleSelector';
@@ -30,13 +29,8 @@ class ProdDetail extends React.Component {
     Promise.all([getCurrentProduct(id), getProductStyles(id), getReviewsMeta(id)])
       .then((results) => {
         const product = results[0].data;
-        const productStyles = results[1].data.results;
-        let defaultStyle;
-        productStyles.forEach((style) => {
-          if (style['default?']) {
-            defaultStyle = style;
-          }
-        });
+        const defaultStyle = this.findDefault(results[1].data.results)[0];
+        const productStyles = this.findDefault(results[1].data.results)[1];
         const prodRating = this.calcRating(results[2].data.ratings);
         this.setState({
           loaded: true,
@@ -55,16 +49,8 @@ class ProdDetail extends React.Component {
       Promise.all([getCurrentProduct(id), getProductStyles(id), getReviewsMeta(id)])
         .then((results) => {
           const product = results[0].data;
-          const productStyles = results[1].data.results;
-          let defaultStyle;
-          productStyles.forEach((style) => {
-            if (style['default?']) {
-              defaultStyle = style;
-            }
-          });
-          if (!defaultStyle) {
-            defaultStyle = productStyles[0];
-          }
+          const defaultStyle = this.findDefault(results[1].data.results)[0];
+          const productStyles = this.findDefault(results[1].data.results)[1];
           const prodRating = this.calcRating(results[2].data.ratings);
           this.setState({
             product,
@@ -78,21 +64,40 @@ class ProdDetail extends React.Component {
   }
 
   handleStyleChange(styleId) {
-    const { selectedStyle, productStyles } = this.state;
-    if (styleId !== selectedStyle.style_id) {
-      productStyles.forEach((style) => {
-        if (styleId === style.style_id) {
-          this.setState({ selectedStyle: style });
-        }
-      });
-    }
+    this.setState({ selectedStyle: styleId });
   }
 
   handleCartAdd(sku, qty) {
-    // eslint-disable-next-line no-alert
     addCart(sku, qty)
       .then(() => console.log('YAY CART ADDED'))
       .catch((err) => console.log(err));
+  }
+
+  styleRepacker(style) {
+    const {
+      name, original_price, sale_price, photos, skus, style_id,
+    } = style;
+    return {
+      [style_id]: {
+        name, original_price, sale_price, photos, skus, style_id,
+      },
+    };
+  }
+
+  findDefault(newStyles) {
+    const styleObj = {};
+    let defaultStyle;
+    newStyles.forEach((style) => {
+      const repackStyle = this.styleRepacker(style);
+      Object.assign(styleObj, repackStyle);
+      if (!defaultStyle) {
+        defaultStyle = style.style_id;
+      }
+      if (style['default?']) {
+        defaultStyle = style.style_id;
+      }
+    });
+    return [defaultStyle, styleObj];
   }
 
   calcRating(ratings) {
@@ -114,8 +119,8 @@ class ProdDetail extends React.Component {
     const { handleReviewScroll } = this.props;
 
     const saleChecker = () => {
-      const { sale_price, original_price } = selectedStyle;
-      if (selectedStyle.sale_price) {
+      const { sale_price, original_price } = productStyles[selectedStyle];
+      if (sale_price) {
         return (
           <span style={{ color: 'red' }}>
             <s style={{ color: 'black' }}>{`$${original_price}`}</s>
@@ -126,36 +131,32 @@ class ProdDetail extends React.Component {
       return <span>{`$${original_price}`}</span>;
     };
 
-    return (
-      <div id="Product Detail">
-        <section className="overview">
-          {selectedStyle.photos
-            && (
+    if (loaded) {
+      return (
+        <div>
+          <section className="overview">
             <ImageGallery
-              selectedStyle={selectedStyle}
+              selectedStyle={productStyles[selectedStyle]}
               currProduct={id}
             />
-            )}
-          <div className="prodSelect">
-            <div className="rating">
-              <div className="prod-stars">
-                {starRating(prodRating)}
+            <div className="prodSelect">
+              <div className="prod-rating">
+                <div className="prod-stars">
+                  {starRating(prodRating)}
+                </div>
+                <span>{`(${prodRating})`}</span>
+                <span
+                  className="reviews-link"
+                  onClick={handleReviewScroll}
+                >
+                  {`See all (${ratingCount}) ratings`}
+                </span>
               </div>
-              <span>{`(${prodRating})`}</span>
-              <span
-                className="reviews-link"
-                onClick={handleReviewScroll}
-              >
-                {`See all (${ratingCount}) ratings`}
-              </span>
-            </div>
-            <span className="prodCategory">{product.category}</span>
-            <span className="productName">{product.name}</span>
-            <div className="price">
-              {saleChecker()}
-            </div>
-            {loaded
-              && (
+              <span className="prodCategory">{product.category}</span>
+              <span className="productName">{product.name}</span>
+              <div className="price">
+                {saleChecker()}
+              </div>
               <StyleSelector
                 currProduct={id}
                 selectedStyle={selectedStyle}
@@ -163,28 +164,35 @@ class ProdDetail extends React.Component {
                 handleStyleChange={this.handleStyleChange}
                 handleCartAdd={this.handleCartAdd}
               />
-              )}
-          </div>
-        </section>
-        <section className="productDetails">
-          <span className="productDescription">
-            {product.description}
-          </span>
-          <div className="vl" />
-          <div className="productFeatures">
-            {!!productStyles.length
-            && product.features.map((feature) => (
-              <span
-                className="productFeature"
-                key={feature.feature}
-              >
-                <b>{feature.feature}</b>
-                {`:     ${feature.value}`}
-              </span>
-            ))}
-          </div>
-        </section>
-      </div>
+            </div>
+          </section>
+          <section className="productDetails">
+            <span className="productDescription">
+              {product.description}
+            </span>
+            <div className="vl" />
+            <div className="productFeatures">
+              {product.features.map((feature) => (
+                <span
+                  className="productFeature"
+                  key={feature.feature}
+                >
+                  <b>{feature.feature}</b>
+                  {`:     ${feature.value}`}
+                </span>
+              ))}
+            </div>
+          </section>
+        </div>
+      );
+    }
+    return (
+      <section className="overview">
+        <img
+          src="https://upload.wikimedia.org/wikipedia/commons/b/b1/Loading_icon.gif"
+          alt="loading"
+        />
+      </section>
     );
   }
 }
